Added several utilities for plotting.
